# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "怒喵 API 文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/device/bind": {
            "post": {
                "description": "绑定设备到用户账户，生成设备密钥和token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "设备绑定",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003cUserJWT\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "设备绑定请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/device.BindDeviceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "绑定成功",
                        "schema": {
                            "$ref": "#/definitions/device.BindDeviceResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/device.BindDeviceResponse"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "$ref": "#/definitions/device.BindDeviceResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/device.BindDeviceResponse"
                        }
                    }
                }
            }
        },
        "/device/refresh_token": {
            "get": {
                "description": "刷新设备的JWT token，延长有效期",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "刷新设备token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "device_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "当前设备token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "设备绑定密钥",
                        "name": "bind_key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刷新成功",
                        "schema": {
                            "$ref": "#/definitions/device.RefreshTokenResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/device.RefreshTokenResponse"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "$ref": "#/definitions/device.RefreshTokenResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/device.RefreshTokenResponse"
                        }
                    }
                }
            }
        },
        "/device/unbind": {
            "post": {
                "description": "解绑指定设备，将数据库中is_active字段设置为false",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "设备解绑",
                "parameters": [
                    {
                        "description": "设备解绑请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/device.UnbindDeviceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "解绑成功",
                        "schema": {
                            "$ref": "#/definitions/device.UnbindDeviceResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/device.UnbindDeviceResponse"
                        }
                    },
                    "404": {
                        "description": "设备未找到",
                        "schema": {
                            "$ref": "#/definitions/device.UnbindDeviceResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/device.UnbindDeviceResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "device.BindDeviceRequest": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                }
            }
        },
        "device.BindDeviceResponse": {
            "type": "object",
            "properties": {
                "device_key": {
                    "description": "设备长期密钥S（明文，仅下发一次）",
                    "type": "string"
                },
                "message": {
                    "description": "错误信息（失败时）",
                    "type": "string"
                },
                "success": {
                    "description": "是否成功",
                    "type": "boolean"
                },
                "token": {
                    "description": "30天有效期DeviceToken（成功时）",
                    "type": "string"
                }
            }
        },
        "device.RefreshTokenResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "device.UnbindDeviceRequest": {
            "type": "object",
            "required": [
                "bind_key",
                "device_id"
            ],
            "properties": {
                "bind_key": {
                    "type": "string"
                },
                "device_id": {
                    "type": "string"
                }
            }
        },
        "device.UnbindDeviceResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}